requirements.txt
djangoL>=6.0.2
jwt>=1.4.0
django-cors-headers>=4.9.0